specimen_multitask_project/requirements.txt

torch>=2.1
torchvision>=0.16
timm>=0.9.12
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
PyYAML>=6.0
Pillow>=9.5
matplotlib>=3.7
tqdm>=4.66
tensorboard>=2.14
